Added test for the users accessing the delete page without login in or accessing a non existent product.

+++ products/views.py
@@ -92,23 +92,31 @@ def product_creator(request):
             
 @login_required
 def product_update(request,product_number):
+    current_user = request.user
     try:
         selected_product = Product.objects.get(pk=product_number)
     except Product.DoesNotExist:
         messages.error(
-                request,
-                "This product does not exist."
-                )
+            request,
+            "This product does not exist."
+            )
         return redirect(inventory)
     else:
         if request.method == "GET":
-            product_form = ProductForm(instance=selected_product)
-            return render(
-                request,
-                "product-form.html",
-                {
-                    "product_form":product_form
-                })
+            if current_user.id is selected_product.seller_id:
+                product_form = ProductForm(instance=selected_product)
+                return render(
+                    request,
+                    "product-form.html",
+                    {
+                        "product_form":product_form
+                    })
+            else:
+                messages.error(
+                    request,
+                    "You are not allowed to update this product's information."
+                    )
+                return redirect(inventory)
         else:
             dirty_product_form = ProductForm(
                 request.POST,
@@ -128,20 +136,29 @@ def product_update(request,product_number):
                     {
                         "product_form":dirty_product_form
                     })
-                
+        
+@login_required        
 def delete_product(request,product_number):
+    current_user = request.user
     try:
-        Product.objects.filter(pk=product_number).delete()
+        selected_product = Product.objects.get(pk=product_number)
     except Product.DoesNotExist:
         messages.error(
                 request,
                 "This product does not exist."
                 )
     else:
-        messages.success(
-            request,
-            "The listing has been successfully deleted!"
-            )
+        if current_user.id is selected_product.seller_id:
+            Product.objects.filter(pk=product_number).delete()
+            messages.success(
+                request,
+                "The listing has been successfully deleted!"
+                )
+        else:
+            messages.error(
+                request,
+                "You are not allowed to delete this product."
+                )
     finally:
         return redirect(inventory)
